requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-crispy-forms==1.11.2
django-import-export==2.5.0
django-localflavor==3.0.1
et-xmlfile==1.0.1
idna==2.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
python-dotenv==0.17.0
python-stdnum==1.16
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
stripe==2.56.0
tablib==3.0.0
typing-extensions==3.7.4.3
urllib3==1.26.4
xlrd==2.0.1
xlwt==1.3.0
